move oss commands from root to subcommands
