+++ docs/content/reference/cli/meshctl_oss.md
@@ -1,0 +1,26 @@
+---
+title: "meshctl oss"
+weight: 5
+---
+## meshctl oss
+
+Install Gloo Mesh
+
+### Options
+
+```
+  -h, --help   help for oss
+```
+
+### Options inherited from parent commands
+
+```
+  -v, --verbose   enable verbose logging
+```
+
+### SEE ALSO
+
+* [meshctl](../meshctl)	 - The Command Line Interface for managing Gloo Mesh.
+* [meshctl oss enterprise](../meshctl_oss_enterprise)	 - Install Gloo Mesh enterprise
+* [meshctl oss install](../meshctl_oss_install)	 - Install Gloo Mesh OSS
+
